db: configuración del db :alien:
